requirement.txt
astroid==1.6.4
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
click==6.7
configparser==3.5.0
decorator==4.3.0
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
ipython==5.7.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.3
opencv-contrib-python==3.4.0.12
pathlib2==2.3.2
pexpect==4.5.0
pickleshare==0.7.4
pkg-resources==0.0.0
Polygon2==2.0.8
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.9.1
scandir==1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
